Add add-beer to root of App
