--- client/src/components/pages/Calendar/AddBeer/index.js
@@ -1,11 +1,0 @@
-import React from "react";
-
-const AddBeer = () => {
-  return (
-    <div>
-      <h1>Add Beer</h1>
-    </div>
-  );
-};
-
-export default AddBeer;
