--- client/src/components/pages/Calendar/Calendar.routes.js
@@ -1,17 +1,0 @@
-import Grid from "./Grid";
-import AddBeer from "./AddBeer";
-
-export default [
-  {
-    path: "/home",
-    rights: "private",
-    component: Grid,
-    exact: true
-  },
-  {
-    path: "/home/add-beer",
-    rights: "private",
-    component: AddBeer,
-    exact: true
-  }
-];
